[Refactor] style 파일 분리
