+++ sidebar/src/components/UI/organisms/bucket_contents/index.js
@@ -1,27 +1,9 @@
 import React from 'react';
-import styled from 'styled-components';
 import { connect } from 'react-redux';
 import { inputDescAction, inputTitleAction } from '../../../../modules/actions/createbucket';
 import BucketDescription from '../../molecules/bucket_decription';
 import BucketTitle from '../../molecules/bucket_title';
-
-const BucketContentsWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-`;
-
-const titleStyle = {
-  width: '100%',
-  marginTop: '10px',
-};
-
-const decriptionStyle = {
-  marginTop: '15px',
-  width: '100%',
-  rows: 4,
-  rowsMax: 10,
-};
+import { BucketContentsWrapper, titleStyle, decriptionStyle } from './style';
 
 const BucketContents = ({
   storeTitle,
