+++ sidebar/src/components/UI/organisms/buttons/style.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const ButtonsWrapper = styled.div`
+display: flex;
+flex-direction: row-reverse;
+margin-top: 15px;
+`;
