+++ sidebar/src/components/templates/user_info_template/index.js
@@ -1,29 +1,8 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
 import UserInfo from '../../UI/organisms/user_info';
 import FollowButton from '../../UI/molecules/follow_button';
 import UserBuckets from '../../UI/organisms/user_buckets';
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-    position: 'relative',
-    padding: theme.spacing(3),
-    transition: theme.transitions.create('margin', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-    color: '#555555',
-  },
-  header: {
-    display: 'flex',
-    alignItems: 'center',
-    padding: theme.spacing(0, 1),
-    ...theme.mixins.toolbar,
-    justifyContent: 'flex-end',
-  },
-}));
-
+import useStyles from './style';
 const UserInfoTemplate = ({ userInfo, buckets, userNo }) => {
   const classes = useStyles();
   return (
